start refactoring import API: sliding, size
