+++ project/plugins.sbt
@@ -1,1 +1,1 @@
-addSbtPlugin("org.hammerlab" % "sbt-parent" % "3.1.1")
+addSbtPlugin("org.hammerlab" % "sbt-parent" % "3.3.0")
