+++ src/main/scala/magic_rdds/package.scala
@@ -1,0 +1,7 @@
+
+import org.hammerlab.magic.rdd
+
+package object magic_rdds {
+  object size extends rdd.size.Size
+  object sliding extends rdd.sliding.ops
+}
