+++ src/main/scala/org/hammerlab/magic/rdd/package.scala
@@ -1,0 +1,4 @@
+package org.hammerlab.magic
+
+package object rdd
+  extends sliding.ops
